Move templates selector to components
