+++ src/app/components/index.ts
@@ -7,3 +7,4 @@ export * from "./vote-line";
 export * from "./action-item";
 export * from "./action-column";
 export * from "./stage-text";
+export * from "./template-selector";
